requirements.txt
# Data Processing
pandas==2.0.3
numpy==1.24.3
scipy==1.11.1

# Machine Learning
scikit-learn==1.3.0
torch==2.0.1
tensorflow==2.13.0
onnxruntime==1.15.1

# Crypto APIs
ccxt==4.0.77
python-binance==1.0.19
websocket-client==1.6.1

# Database
supabase==1.0.4
psycopg2-binary==2.9.7

# Visualization & Dashboard
streamlit==1.25.0
plotly==5.15.0
matplotlib==3.7.2
seaborn==0.12.2

# Sentiment Analysis
transformers==4.32.1
vaderSentiment==3.3.2
requests==2.31.0

# Technical Indicators
ta-lib==0.4.27
pandas-ta==0.3.14b0

# Utilities
python-dotenv==1.0.0
asyncio
aiohttp==3.8.5
schedule==1.2.0
EOF

